segmented audio support
if a story comes back with an array of urls, we will play them in order,
one at a time.

`getCurrentSegment` either returns the audio field (if there it's just
a string) or the value in the current index on the story model's audio
field, as tracked by a private attr

`playNextSegment` is fired in the `audio-ended` handler only if the
current story has segmented audio. That line isn't great, because it is
forced to return early so the bumper doesn't play. that block was
cleaned up in another branch, thus the FIXME.

this branch also relies on an update to `ember-hifi` to support passd in
argumets to the `play` methods, so we cna assure that he played piece
will start from a specific position.