region -> playContext
playContext is what this value is used for. region is an old analytics
thing we're getting rid of.

also the value can be specified by its context in a few places, either
in the code or else passed in, so we can get rid of a computed and
otherwise set the value.